backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10
beautifulsoup4==4.12.2
PyGithub==2.1.1
gitpython==3.1.40
